refactor structure names in Stake and unStake
